+++ src/ska_low_mccs_spshw/station/station_device.py
@@ -102,6 +102,8 @@ class SpsStation(MccsBaseDevice, SKAObsDevice):
 
     LMCDaqTRL = device_property(dtype=str, default_value="")
     BandpassDaqTRL = device_property(dtype=str, default_value="")
+    WRENTRL = device_property(dtype=str, default_value="")
+
     AntennaConfigURI = device_property(
         dtype=(str,),
         default_value=[],
@@ -172,6 +174,7 @@ class SpsStation(MccsBaseDevice, SKAObsDevice):
             f"\tTileFQDNs: {self.TileFQDNs}\n"
             f"\tLMCDaqTRL: {self.LMCDaqTRL}\n"
             f"\tBandpassDaqTRL: {self.BandpassDaqTRL}\n"
+            f"\tWRENTRL: {self.WRENTRL}\n"
             f"\tSubrackFQDNs: {self.SubrackFQDNs}\n"
             f"\tSdnFirstInterface: {self.SdnFirstInterface}\n"
             f"\tSdnGateway: {self.SdnGateway}\n"
@@ -255,6 +258,7 @@ class SpsStation(MccsBaseDevice, SKAObsDevice):
             # It appears the test context inputs a space into empty strings.
             self.LMCDaqTRL if self.LMCDaqTRL != " " else "",
             self.BandpassDaqTRL if self.BandpassDaqTRL != " " else "",
+            self.WRENTRL if self.WRENTRL != " " else "",
             ipaddress.IPv4Interface(self.SdnFirstInterface),
             ipaddress.IPv4Address(self.SdnGateway) if self.SdnGateway else None,
             ipaddress.IPv4Address(self.CspIngestIp) if self.CspIngestIp else None,
@@ -848,6 +852,25 @@ class SpsStation(MccsBaseDevice, SKAObsDevice):
         """
         self.BandpassDaqTRL = value
         self.component_manager._bandpass_daq_trl = value
+
+    @attribute(dtype=str)
+    def WrenTRL(self: SpsStation) -> str:
+        """
+        Report the Tango Resource Locator for this SpsStation's WREN instance.
+
+        :return: Return the current WREN TRL.
+        """
+        return self.WRENTRL
+
+    @WrenTRL.write  # type: ignore[no-redef]
+    def WrenTRL(self: SpsStation, value: str) -> None:
+        """
+        Set the Tango Resource Locator for this SpsStation's WREN instance.
+
+        :param value: The new WREN TRL.
+        """
+        self.WRENTRL = value
+        self.component_manager._wren_trl = value
 
     @attribute(dtype="DevBoolean")
     def isCalibrated(self: SpsStation) -> bool: