+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [THORN-125] Improve observability when UNPROGRAMMED.
 * [THORN-125] Bump daq version: 6.1.0 -> 6.1.1
+* [THORN-611] Add `WrenTRL` device property to SpsStation for WREN TRL
 
 ## 14.0.1